requirements.txt
certifi==2019.11.28
mkl-fft==1.0.15
mkl-service==2.3.0
numpy==1.18.1
olefile==0.46
Pillow==7.0.0
six==1.13.0
